requirement.txt
Django==4.2.5
djangorestframework==3.14.0
mysqlclient==2.2.0
